#71 UI: extract routes to separate file to find routes easier
