+++ client/app.js
@@ -1,6 +1,7 @@
 import overallProgressComponent from './overall-progress.component';
 import todaysProgressComponent from './todays-progress.component';
 import navBarComponent from './nav-bar.component';
+import routes from './routes.js';
 
 angular
     .module('bookmaster', [
@@ -10,20 +11,4 @@ angular
         todaysProgressComponent.default,
         navBarComponent.default
     ])
-    .config(Routes);
-
-Routes.$inject = ['$stateProvider', '$urlRouterProvider'];
-function Routes($stateProvider, $urlRouterProvider) {
-    $stateProvider
-        .state({
-            name: 'overall',
-            url: '/overall',
-            component: overallProgressComponent.default
-        })
-        .state({
-            name: 'today',
-            url: '/today',
-            component: todaysProgressComponent.default
-        });
-    $urlRouterProvider.otherwise('/overall');
-}
+    .config(routes.default);
